requirements.txt
gradio==4.3.0
python-dotenv==1.0.0
replicate==0.18.1
